+++ src/gui_button.h
@@ -12,6 +12,6 @@
 
 #include "gui.h"
 
-uint8_t ButtonWidget(uint8_t id, uint8_t x, uint8_t y, uint8_t w, uint8_t h, uint8_t* text);
+uint8_t ButtonWidget(uint8_t id, uint16_t x, uint16_t y, uint16_t w, uint16_t h, uint8_t* text);
 
 #endif
